`\LastDragon_ru\LaraASP\Documentator\Markdown\Mutator\Mutation` moved to `\LastDragon_ru\LaraASP\Documentator\Markdown\Contracts\Mutation`.
