+++ packages/documentator/src/Markdown/Contracts/Mutation.php
@@ -1,6 +1,6 @@
 <?php declare(strict_types = 1);
 
-namespace LastDragon_ru\LaraASP\Documentator\Markdown\Mutator;
+namespace LastDragon_ru\LaraASP\Documentator\Markdown\Contracts;
 
 use LastDragon_ru\LaraASP\Documentator\Markdown\Document;
 use LastDragon_ru\LaraASP\Documentator\Markdown\Mutator\Mutagens\Delete;
